feat(api): add ds-api-types crate
