+++ packages/ds-api-types/src/lib.rs
@@ -1,0 +1,2 @@
+mod entropic_packument;
+pub use entropic_packument::*;
