+++ packages/pick-version/tests/tests.rs
@@ -1,20 +1,21 @@
+use ds_api_types::EntropicPackument;
 use maplit::hashmap;
 use parse_package_arg::PackageArg;
 use semver::{Version, VersionReq};
 use ssri::Integrity;
 
-use pick_version::{Packument, Picker};
+use pick_version::Picker;
 
 #[test]
 fn basic_carat_range() {
-    let packument = Packument {
+    let packument = EntropicPackument {
         versions: hashmap! {
             Version::parse("1.0.0").unwrap() => "sha256-100".parse().unwrap(),
             Version::parse("1.0.1").unwrap() => "sha256-101".parse().unwrap(),
             Version::parse("1.0.2").unwrap() => "sha256-102".parse().unwrap(),
             Version::parse("2.0.0").unwrap() => "sha256-200".parse().unwrap(),
         },
-        ..Packument::default()
+        ..EntropicPackument::default()
     };
     let sri = Picker::new()
         .pick(
@@ -31,14 +32,14 @@ fn basic_carat_range() {
 
 #[test]
 fn basic_tilde_range() {
-    let packument = Packument {
+    let packument = EntropicPackument {
         versions: hashmap! {
             Version::parse("1.0.0").unwrap() => "sha256-100".parse().unwrap(),
             Version::parse("1.0.1").unwrap() => "sha256-101".parse().unwrap(),
             Version::parse("1.0.2").unwrap() => "sha256-102".parse().unwrap(),
             Version::parse("2.0.0").unwrap() => "sha256-200".parse().unwrap(),
         },
-        ..Packument::default()
+        ..EntropicPackument::default()
     };
     let sri = Picker::new()
         .pick(
@@ -55,14 +56,14 @@ fn basic_tilde_range() {
 
 #[test]
 fn basic_math_range() {
-    let packument = Packument {
+    let packument = EntropicPackument {
         versions: hashmap! {
             Version::parse("1.0.0").unwrap() => "sha256-100".parse().unwrap(),
             Version::parse("1.0.1").unwrap() => "sha256-101".parse().unwrap(),
             Version::parse("1.0.2").unwrap() => "sha256-102".parse().unwrap(),
             Version::parse("2.0.0").unwrap() => "sha256-200".parse().unwrap(),
         },
-        ..Packument::default()
+        ..EntropicPackument::default()
     };
     let sri = Picker::new()
         .pick(
@@ -79,14 +80,14 @@ fn basic_math_range() {
 
 #[test]
 fn basic_version_match() {
-    let packument = Packument {
+    let packument = EntropicPackument {
         versions: hashmap! {
             Version::parse("1.0.0").unwrap() => "sha256-100".parse().unwrap(),
             Version::parse("1.0.1").unwrap() => "sha256-101".parse().unwrap(),
             Version::parse("1.0.2").unwrap() => "sha256-102".parse().unwrap(),
             Version::parse("2.0.0").unwrap() => "sha256-200".parse().unwrap(),
         },
-        ..Packument::default()
+        ..EntropicPackument::default()
     };
     let sri = Picker::new()
         .pick(
@@ -103,7 +104,7 @@ fn basic_version_match() {
 
 #[test]
 fn basic_tag_match() {
-    let packument = Packument {
+    let packument = EntropicPackument {
         tags: hashmap! {
             "latest".into() => Version::parse("1.0.1").unwrap()
         },
@@ -113,7 +114,7 @@ fn basic_tag_match() {
             Version::parse("1.0.2").unwrap() => "sha256-102".parse().unwrap(),
             Version::parse("2.0.0").unwrap() => "sha256-200".parse().unwrap(),
         },
-        ..Packument::default()
+        ..EntropicPackument::default()
     };
     let sri = Picker::new()
         .pick(
@@ -130,7 +131,7 @@ fn basic_tag_match() {
 
 #[test]
 fn tag_match_with_custom_default_tag() {
-    let packument = Packument {
+    let packument = EntropicPackument {
         tags: hashmap! {
             "something".into() => Version::parse("1.0.1").unwrap(),
             "latest".into() => Version::parse("1.0.2").unwrap()
@@ -141,7 +142,7 @@ fn tag_match_with_custom_default_tag() {
             Version::parse("1.0.2").unwrap() => "sha256-102".parse().unwrap(),
             Version::parse("2.0.0").unwrap() => "sha256-200".parse().unwrap(),
         },
-        ..Packument::default()
+        ..EntropicPackument::default()
     };
     let sri = Picker::new()
         .default_tag("something".into())
@@ -159,7 +160,7 @@ fn tag_match_with_custom_default_tag() {
 
 #[test]
 fn star_range_uses_default_tag() {
-    let packument = Packument {
+    let packument = EntropicPackument {
         tags: hashmap! {
             "latest".into() => Version::parse("1.0.0-pre.0").unwrap(),
             "beta".into() => Version::parse("2.0.0-beta.0").unwrap(),
@@ -170,7 +171,7 @@ fn star_range_uses_default_tag() {
             Version::parse("2.0.0-beta.0").unwrap() => "sha256-200b0".parse().unwrap(),
             Version::parse("2.0.0-beta.1").unwrap() => "sha256-200b1".parse().unwrap(),
         },
-        ..Packument::default()
+        ..EntropicPackument::default()
     };
     let sri = Picker::new()
         .default_tag("beta".into())
@@ -199,14 +200,14 @@ fn star_range_uses_default_tag() {
 
 #[test]
 fn error_if_no_match() {
-    let packument = Packument {
+    let packument = EntropicPackument {
         versions: hashmap! {
             Version::parse("1.0.0").unwrap() => "sha256-100".parse().unwrap(),
             Version::parse("1.0.1").unwrap() => "sha256-101".parse().unwrap(),
             Version::parse("1.0.2").unwrap() => "sha256-102".parse().unwrap(),
             Version::parse("2.0.0").unwrap() => "sha256-200".parse().unwrap(),
         },
-        ..Packument::default()
+        ..EntropicPackument::default()
     };
     let sri = Picker::new().pick(
         &packument,
@@ -221,7 +222,7 @@ fn error_if_no_match() {
 
 #[test]
 fn error_if_no_versions() {
-    let packument = Packument::default();
+    let packument = EntropicPackument::default();
     let sri = Picker::new().pick(
         &packument,
         &PackageArg::Range {
